Move CliSpec's to to their respective package
